backend/requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
iso8601==1.0.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
uWSGI>=2.0.18,<2.1
psycopg2-binary
# backports.zoneinfo==0.2.1
